+++ pkg/http/ui/form_create.go
@@ -19,7 +19,7 @@ func (h *Handler) handleFormCreate(ctx context.Context, path string, userID, use
 
 	obj := h.svc.New(path)
 
-	now := uint64(time.Now().UTC().Unix())
+	now := time.Now().UTC().Unix()
 	userIDInt, err := strconv.ParseUint(userID, 10, 64)
 	if err != nil {
 		errCode := logger.LogError("error converting user id to int", logAttrHandler, logAttrPath, logger.AttrError(err))
